fix(change desc): change interface of web
transform pages
