+++ src/components/steps/signup/Step2_signupForm/index.tsx
@@ -1,4 +1,11 @@
-import { FormControlLabel, Radio, RadioGroup, TextField } from '@mui/material';
+import {
+  Checkbox,
+  FormControl,
+  FormControlLabel,
+  FormGroup,
+  FormHelperText,
+  TextField,
+} from '@mui/material';
 import { motion } from 'framer-motion';
 import { useAtom } from 'jotai';
 import React from 'react';
@@ -7,37 +14,43 @@ import { useTranslation } from 'react-i18next';
 
 import { multiStepFormAtom } from '../signup_state';
 
-interface Step2Data {
-  organization: string;
-  duration: string;
-  legalWorkType: string;
-  interestArea: string;
-  specificSkills: string;
-  practicalExperience: string;
-}
-
-interface Step2Props {
+interface Step4Props {
   next: () => void;
   back: () => void;
 }
 
-const Step2: React.FC<Step2Props> = ({ next, back }) => {
+interface FormData {
+  caseType: string;
+  articleNumber: string;
+  articleText: string;
+  issuingAuthority: string;
+  judgmentSummary: string;
+  keyHighlights: string;
+  practicalExamples: string;
+  priorityLegalAreas: string;
+}
+const Step2: React.FC<Step4Props> = ({ next, back }) => {
   const { t } = useTranslation();
-  const [formData, setFormData] = useAtom(multiStepFormAtom);
+  const [formData] = useAtom(multiStepFormAtom);
 
   const {
     control,
     handleSubmit,
     formState: { errors },
-  } = useForm<Step2Data>({
-    defaultValues: formData.step2,
+  } = useForm<FormData>({
+    defaultValues: formData.step4 || {
+      caseType: '',
+      articleNumber: '',
+      articleText: '',
+      issuingAuthority: '',
+      judgmentSummary: '',
+      keyHighlights: '',
+      practicalExamples: '',
+      priorityLegalAreas: '',
+    },
   });
 
-  const onSubmit = (data: Step2Data) => {
-    setFormData((prev) => ({
-      ...prev,
-      step2: data,
-    }));
+  const onSubmit = () => {
     next();
   };
 
@@ -49,176 +62,284 @@ const Step2: React.FC<Step2Props> = ({ next, back }) => {
         animate={{ opacity: 1, y: 0 }}
         transition={{ duration: 0.5 }}
       >
-        <div className="mt-10 flex w-full flex-col gap-2 border-b p-5 md:px-10">
-          <h1 className="text-lg font-[500]">{t('Organization')}</h1>
+        <div className="mt-10 flex flex-col gap-2 border-b p-5 md:px-10">
+          <h1 className="text-lg font-[500]">{t('Case Type')}</h1>
           <p className="text-xs text-gray-400">
             {t('Please provide the information required in the below form')}
           </p>
         </div>
 
-        <div className="flex w-full flex-col gap-10 p-5 md:p-10 lg:w-[80%]">
-          <div className="flex flex-wrap justify-between gap-5 lg:gap-10">
+        <div className="flex w-full flex-col justify-between gap-5 p-5 text-sm md:p-10 lg:w-[80%] lg:gap-10">
+          <div className="flex flex-wrap justify-between">
             <Controller
-              name="organization"
+              name="caseType"
               control={control}
               rules={{ required: t('This field is required') }}
               render={({ field }) => (
-                <TextField
-                  placeholder={t('Organization')}
-                  variant="outlined"
+                <FormControl
                   className="w-[45%]"
-                  {...field}
-                  error={!!errors.organization}
-                  helperText={errors.organization?.message}
-                />
+                  error={Boolean(errors.caseType)}
+                >
+                  <TextField
+                    {...field}
+                    placeholder={t('Case Type')}
+                    variant="outlined"
+                    value={field.value || ''}
+                  />
+                  {errors.caseType && (
+                    <FormHelperText>{errors.caseType.message}</FormHelperText>
+                  )}
+                </FormControl>
               )}
             />
+
             <Controller
-              name="duration"
+              name="articleNumber"
               control={control}
               rules={{ required: t('This field is required') }}
               render={({ field }) => (
-                <TextField
-                  placeholder={t('Duration')}
-                  variant="outlined"
+                <FormControl
                   className="w-[45%]"
-                  {...field}
-                  error={!!errors.duration}
-                  helperText={errors.duration?.message}
-                />
+                  error={Boolean(errors.articleNumber)}
+                >
+                  <TextField
+                    {...field}
+                    placeholder={t('Article Number')}
+                    variant="outlined"
+                    value={field.value || ''}
+                  />
+                  {errors.articleNumber && (
+                    <FormHelperText>
+                      {errors.articleNumber.message}
+                    </FormHelperText>
+                  )}
+                </FormControl>
               )}
             />
           </div>
 
-          <div className="flex">
-            <Controller
-              name="legalWorkType"
-              control={control}
-              rules={{ required: t('This field is required') }}
-              render={({ field }) => (
+          <Controller
+            name="articleText"
+            control={control}
+            rules={{ required: t('This field is required') }}
+            render={({ field }) => (
+              <FormControl
+                error={Boolean(errors.articleText)}
+                className="w-full"
+              >
                 <TextField
-                  placeholder={t('Type of Legal Work')}
-                  variant="outlined"
-                  className="w-full lg:w-[100%]"
                   {...field}
-                  error={!!errors.legalWorkType}
-                  helperText={errors.legalWorkType?.message}
+                  placeholder={t('Text of the Article or Judgment')}
+                  variant="outlined"
+                  value={field.value || ''}
                 />
-              )}
-            />
-          </div>
+                {errors.articleText && (
+                  <FormHelperText>{errors.articleText.message}</FormHelperText>
+                )}
+              </FormControl>
+            )}
+          />
 
-          <div>
-            <Controller
-              name="interestArea"
-              control={control}
-              rules={{ required: t('This field is required') }}
-              render={({ field }) => (
+          <Controller
+            name="issuingAuthority"
+            control={control}
+            rules={{ required: t('This field is required') }}
+            render={({ field }) => (
+              <FormControl
+                error={Boolean(errors.issuingAuthority)}
+                className="w-full"
+              >
                 <TextField
-                  placeholder={t('Which area of law interests you the most')}
-                  variant="outlined"
-                  className="w-full lg:w-[100%]"
                   {...field}
-                  error={!!errors.interestArea}
-                  helperText={errors.interestArea?.message}
+                  placeholder={t('Issuing Authority')}
+                  variant="outlined"
+                  value={field.value || ''}
                 />
-              )}
-            />
-          </div>
-
-          <div className="flex flex-col gap-3">
-            <h2 className="text-sm font-[500]">
-              {t(
-                'Do you have specific interests in developing particular skills such as drafting legal memos, case management, or advocacy ?',
-              )}
-            </h2>
-            <Controller
-              name="specificSkills"
-              control={control}
-              rules={{ required: t('This field is required') }}
-              render={({ field }) => (
-                <RadioGroup {...field} className="text-gray-700 ">
-                  <FormControlLabel
-                    value="Yes"
-                    control={
-                      <Radio
-                        sx={{
-                          color: 'black',
-                          '&.Mui-checked': { color: 'black' },
-                        }}
-                      />
-                    }
-                    label={t('Yes')}
-                  />
-                  <FormControlLabel
-                    value="No"
-                    control={
-                      <Radio
-                        sx={{
-                          color: 'black',
-                          '&.Mui-checked': { color: 'black' },
-                        }}
-                      />
-                    }
-                    label={t('No')}
-                  />
-                </RadioGroup>
-              )}
-            />
-            {errors.specificSkills && (
-              <p className="text-xs text-red-500">
-                {errors.specificSkills.message}
-              </p>
+                {errors.issuingAuthority && (
+                  <FormHelperText>
+                    {errors.issuingAuthority.message}
+                  </FormHelperText>
+                )}
+              </FormControl>
             )}
-          </div>
+          />
 
-          <div className="flex flex-col gap-3">
-            <h2 className="text-sm font-[500]">
-              {t('Do you have practical experience in the legal field?')}
-            </h2>
-            <Controller
-              name="practicalExperience"
-              control={control}
-              rules={{ required: t('This field is required') }}
-              render={({ field }) => (
-                <RadioGroup {...field} className="text-gray-700 ">
-                  <FormControlLabel
-                    value="Yes"
-                    control={
-                      <Radio
-                        sx={{
-                          color: 'black',
-                          '&.Mui-checked': { color: 'black' },
-                        }}
+          <div className="flex flex-col gap-10">
+            <div className="flex flex-col gap-2">
+              <h1 className="text-lg font-[500]">{t('Judgment Summary')}</h1>
+              <Controller
+                name="judgmentSummary"
+                control={control}
+                rules={{ required: t('This field is required') }}
+                render={({ field }) => (
+                  <FormControl
+                    error={Boolean(errors.judgmentSummary)}
+                    className="w-full"
+                  >
+                    <TextField
+                      {...field}
+                      variant="outlined"
+                      placeholder={t('Judgment summary is required')}
+                      value={field.value || ''}
+                    />
+                    {errors.judgmentSummary && (
+                      <FormHelperText>
+                        {errors.judgmentSummary.message}
+                      </FormHelperText>
+                    )}
+                  </FormControl>
+                )}
+              />
+            </div>
+
+            <div className="flex flex-col gap-2">
+              <h1 className="mt-5 font-[500] text-black">
+                {t('Key Highlights')}
+              </h1>
+              <Controller
+                name="keyHighlights"
+                control={control}
+                rules={{ required: t('This field is required') }}
+                render={({ field }) => (
+                  <FormControl
+                    error={Boolean(errors.keyHighlights)}
+                    className="w-full"
+                  >
+                    <TextField
+                      {...field}
+                      variant="outlined"
+                      placeholder={t('Key Highlights')}
+                      value={field.value || ''}
+                    />
+                    {errors.keyHighlights && (
+                      <FormHelperText>
+                        {errors.keyHighlights.message}
+                      </FormHelperText>
+                    )}
+                  </FormControl>
+                )}
+              />
+            </div>
+
+            <div className="flex flex-col gap-2">
+              <h1 className="mt-5 font-[500] text-black">
+                {t('Practical Examples of the Judgment')}
+              </h1>
+              <Controller
+                name="practicalExamples"
+                control={control}
+                rules={{ required: t('This field is required') }}
+                render={({ field }) => (
+                  <FormControl
+                    error={Boolean(errors.practicalExamples)}
+                    className="w-full"
+                  >
+                    <TextField
+                      {...field}
+                      variant="outlined"
+                      placeholder={t('Practical Examples of the Judgment')}
+                      value={field.value || ''}
+                    />
+                    {errors.practicalExamples && (
+                      <FormHelperText>
+                        {errors.practicalExamples.message}
+                      </FormHelperText>
+                    )}
+                  </FormControl>
+                )}
+              />
+            </div>
+
+            <div className="flex flex-col gap-2">
+              <h1 className="mt-5 font-[500] text-black">
+                {t('Priority Legal Areas for the Application')}
+              </h1>
+              <Controller
+                name="priorityLegalAreas"
+                control={control}
+                rules={{ required: t('This field is required') }}
+                render={({ field }) => (
+                  <FormControl
+                    error={Boolean(errors.priorityLegalAreas)}
+                    className="w-full"
+                  >
+                    <FormGroup className="text-gray-700">
+                      <FormControlLabel
+                        control={
+                          <Checkbox
+                            {...field}
+                            value="Labor and Corporate Laws"
+                            sx={{
+                              color: 'black',
+                              '&.Mui-checked': {
+                                color: 'black',
+                              },
+                            }}
+                          />
+                        }
+                        label={t('Labor and Corporate Laws')}
                       />
-                    }
-                    label={t('Yes')}
-                  />
-                  <FormControlLabel
-                    value="No"
-                    control={
-                      <Radio
-                        sx={{
-                          color: 'black',
-                          '&.Mui-checked': { color: 'black' },
-                        }}
+                      <FormControlLabel
+                        control={
+                          <Checkbox
+                            {...field}
+                            value="Family Laws (Marriage, Divorce, Custody)"
+                            sx={{
+                              color: 'black',
+                              '&.Mui-checked': {
+                                color: 'black',
+                              },
+                            }}
+                          />
+                        }
+                        label={t('Family Laws (Marriage, Divorce, Custody)')}
                       />
-                    }
-                    label={t('No')}
-                  />
-                </RadioGroup>
-              )}
-            />
-            {errors.practicalExperience && (
-              <p className="text-xs text-red-500">
-                {errors.practicalExperience.message}
-              </p>
-            )}
+                      <FormControlLabel
+                        control={
+                          <Checkbox
+                            {...field}
+                            value="International Laws"
+                            sx={{
+                              color: 'black',
+                              '&.Mui-checked': {
+                                color: 'black',
+                              },
+                            }}
+                          />
+                        }
+                        label={t('International Laws')}
+                      />
+                      <FormControlLabel
+                        control={
+                          <Checkbox
+                            {...field}
+                            value="Commercial Laws (Contracts, Intellectual Property)"
+                            sx={{
+                              color: 'black',
+                              '&.Mui-checked': {
+                                color: 'black',
+                              },
+                            }}
+                          />
+                        }
+                        label={t(
+                          'Commercial Laws (Contracts, Intellectual Property)',
+                        )}
+                      />
+                    </FormGroup>
+                    {errors.priorityLegalAreas && (
+                      <FormHelperText>
+                        {errors.priorityLegalAreas.message}
+                      </FormHelperText>
+                    )}
+                  </FormControl>
+                )}
+              />
+            </div>
           </div>
         </div>
       </motion.div>
-
-      <div className="mb-10 mt-8 flex justify-between px-5 md:px-10">
+      <div className="my-10 flex justify-between px-5 md:px-10">
         <motion.button
           whileHover={{
             scale: 1.05,
