+++ src/components/steps/signup/state.ts
@@ -4,19 +4,20 @@ export const multiStepFormAtom = atom({
   step1: {
     fullName: '',
     email: '',
-    mobileNumber: '',
     institution: '',
     fieldOfStudy: '',
     academicLevel: '',
     legalTraining: '',
   },
   step2: {
-    organization: '',
-    duration: '',
-    legalWorkType: '',
-    interestArea: '',
-    specificSkills: '',
-    practicalExperience: '',
+    caseType: '',
+    articleNumber: '',
+    articleText: '',
+    issuingAuthority: '',
+    judgmentSummary: '',
+    keyHighlights: '',
+    practicalExamples: '',
+    priorityLegalAreas: '',
   },
   step3: {
     familiarityLevel: '',
@@ -25,14 +26,12 @@ export const multiStepFormAtom = atom({
     careerGoals: '',
   },
   step4: {
-    lawName: '',
-    articleNumber: '',
-    articleText: '',
-    issuingAuthority: '',
-    judgmentSummary: '',
-    keyHighlights: '',
-    practicalExamples: '',
-    priorityLegalAreas: '',
+    LawFirm: '',
+    Mobile_Number: '',
+    legalWorkType: '',
+    interestArea: '',
+    specificSkills: '',
+    practicalExperience: '',
   },
   step5: {},
 });
